Uploaded R script to simplify template generation
Use this script to create a csv file that can be used as input for the interpretablePA R-package. The interpretablePA package will process this csv to produce another csv file that allows you to compare your accelerometer data (average acceleration and intensity gradient) against our reference values, which are specific to age and sex. 

The output will be presented in terms of percentages of the predicted values and Z-scores. These metrics are useful for assessing whether an individual's physical activity level is within the expected range compared to individuals of similar age and sex in the general population. After running the full script, the interpretablePA Shiny app will be run. User should then go to '1) User data' and 'Cohort-level data (raw)' and upload the csv file generated above. The csv file will be stored in the working directory as 'interpretablePA_template_DATE.csv'.